large blobs will deadlock gpg
pipe buffers seem to be limited to 64k bytes, after that gpg will block
waiting for gpgmeh to read from the pipe. On every attempted read,
gpgmeh should read from all pipes so gpg doesn't deadlock.

Deadlock == gpg is waiting to write to stdout, while gpgmeh is waiting
to read on the status-fd